Following guidelines C.46 and C.164: make all conversions explicit.
It revealed a bug in Mpz::pow_mod

+++ examples/CL_HSM2k_setup.cpp
@@ -80,7 +80,7 @@ main (int argc, char *argv[])
       try
       {
         BICYCL::Mpz kmpz(argv[1]);
-        k = kmpz;
+        k = static_cast<unsigned long> (kmpz);
       }
       catch (const std::runtime_error& e)
       {
